// src/app/models/usuario.ts
export class Usuario {
    CodigoUsuario: number;
    Email: string;
    Nome: string;
    SobreNome: string;
    Telefone: any;
    Celular: any;
    Senha: any;
    ReSenha: any;
    CpfCnpj: any;
    Matricula: string;
    Endereco: string;
    Bloqueio: boolean;
    Cep: any;
    EnderecoRua: string;
    EnderecoNumero: string;
    Bairro: any;
    Cidade: any;
    CodigoBaseUF: any;
    CodigoBaseMunicipio: any;        
    NomeSocial:string;
    Municipios: any [];  
    UFs: any [];  
    CpfCnpjEmpresa: string;         
    CodigoUnidade:Number;
    Unidade:string;
    Unidades: any [];  
    VwAvatarPath:any;   
}